# Conflicts:
#	app/src/main/java/manager/PhotoManager.java
#	app/src/main/java/manager/UserAtPhotoManager.java

+++ app/src/main/java/Entities/Photo.java
@@ -8,12 +8,12 @@ import java.util.Date;
 
 public class Photo {
     private int id;
-    private float lat;
-    private float lon;
+    private double lat;
+    private double lon;
     private Date date;
     private String src;
 
-    public Photo(int id, float lat, float lon, Date date, String src) {
+    public Photo(int id, double lat, double lon, Date date, String src) {
         this.id = id;
         this.lat = lat;
         this.lon = lon;
@@ -33,7 +33,7 @@ public class Photo {
         this.id = id;
     }
 
-    public float getLat() {
+    public double getLat() {
         return lat;
     }
 
@@ -41,7 +41,7 @@ public class Photo {
         this.lat = lat;
     }
 
-    public float getLon() {
+    public double getLon() {
         return lon;
     }
